Assignment Lab5 complete
Assignment:

We need to make ViewFrame class an observer of the database.

The two buttons checkIn and checkOut do the check in or check out job with the help of database controller.

There is also a panel on the ViewFrame which shows the result of pressing on of the two buttons.